+++ CHANGELOG.md
@@ -35,6 +35,8 @@
 
 * Define a current user and current instance with lamin-cli prior to testing and generating documentation in the CI (PR #23).
 
+* Add a simple unit test which queries laminlabs/lamindata (PR #27).
+
 ## BUG FIXES
 
 * Fixed the parsing of the env files in `~/.lamin` due to changes in the lamindb-setup Python package (PR #12).